refactor : 부적절한 폴더 네이밍 SortButtonGroup -> FolderControls 로 변경
